refactor: defer extension provider verification until SDK release
