+++ .github/workflows/lint-ext-azure-ai-connections.yml
@@ -5,7 +5,6 @@ on:
     paths:
       - "cli/azd/extensions/azure.ai.connections/**"
       - ".github/workflows/lint-ext-azure-ai-connections.yml"
-      - ".github/workflows/verify-ext-providers.yml"
     branches: [main]
 
 concurrency:
@@ -19,10 +18,5 @@ permissions:
 jobs:
   lint:
     uses: ./.github/workflows/lint-go.yml
-    with:
-      working-directory: cli/azd/extensions/azure.ai.connections
-
-  verify-providers:
-    uses: ./.github/workflows/verify-ext-providers.yml
     with:
       working-directory: cli/azd/extensions/azure.ai.connections
